Extract NetAddr tests for traits into a matching submodule tree
Instead of having very long files with lots of indepedent behavior, we
can test the code in the same file it is written, and this proves to be
a much better way of organizing things anyhow.

In effect, each impl for a given trait gets a file, and all of the
pertinent behavior is tested in the same file.  This means that there
should be very few external "use" statements within the tests mod,
except in cases where an impl is expected to use another trait or impl
somewhere.

Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
